requirements.txt
# core bot functionality
python-telegram-bot>=20.0
APScheduler>=3.9

# for parsing dates flexibly
dateparser>=1.2.2

# load .env
python-dotenv>=1.0

# transitively required by dateparser
python-dateutil>=2.9
pytz>=2025.2
regex>=2024.11.6
six>=1.17.0
tzlocal>=5.3.1

# watchdog for auto restart while code is running
watchgod>=0.8